Write permissions as UInt16 instead of byte (too many permissions to write as byte anymore)

+++ Barotrauma/BarotraumaShared/Source/Networking/GameServer.cs
@@ -826,7 +826,7 @@ namespace Barotrauma.Networking
             ClientPermissions command = ClientPermissions.None;
             try
             {
-                command = (ClientPermissions)inc.ReadByte();
+                command = (ClientPermissions)inc.ReadUInt16();
             }
 
             catch
